+++ ci/install-from-pypi.sh
@@ -2,6 +2,6 @@
 conda update --all -q
 conda create -n linearmodels-test python
 source activate linearmodels-test
-pip install numpy scipy pandas matplotlib matplotlib seaborn
+pip install numpy scipy pandas matplotlib matplotlib seaborn cython
 pip install statsmodels
 if [[ ! -z ${XARRAY} ]]; then pip install xarray==${XARRAY}; fi