+++ doc/source/changes.rst
@@ -3,6 +3,7 @@ Change Log
 
 Since Version 4.12
 ==================
+* Added optional Cythonized node selection for dropping singletons
 * Added preconditioning to the dummy variable matrix when ``use_lsmr=True``
   in :func:`~linearmodels.panel.model.PanelOLS.fit`. In models with many
   effects, this can reduce run time by a factor of 4 or more.